tidy up icons
