--- app/arrowDown.tsx
@@ -1,19 +1,0 @@
-export const ArrowDown = ({ color, ...props }: any) => {
-  return (
-    <svg
-      width="15"
-      height="9"
-      viewBox="0 0 15 9"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M1 1L7.15 7.15C7.19489 7.19783 7.24911 7.23595 7.30931 7.26201C7.3695 7.28807 7.4344 7.30152 7.5 7.30152C7.5656 7.30152 7.6305 7.28807 7.69069 7.26201C7.75089 7.23595 7.80511 7.19783 7.85 7.15L14 1"
-        stroke="white"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  );
-};
