--- app/arrowUp.tsx
@@ -1,19 +1,0 @@
-export const ArrowUp = ({ color, ...props }: any) => {
-  return (
-    <svg
-      width="15"
-      height="9"
-      viewBox="0 0 15 9"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M1 7.30176L7.15 1.15176C7.19489 1.10393 7.24911 1.06581 7.30931 1.03974C7.3695 1.01368 7.4344 1.00024 7.5 1.00024C7.5656 1.00024 7.6305 1.01368 7.69069 1.03974C7.75089 1.06581 7.80511 1.10393 7.85 1.15176L14 7.30176"
-        stroke="white"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  );
-};
