--- app/assets/Bell.tsx
@@ -1,14 +1,0 @@
-export const Bell = ({ color }: { color?: string }) => (
-  <svg
-    width="100%"
-    height="100%"
-    viewBox="0 0 22 21"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M19.1881 16.4137V17.3629H1.91883V16.4137L3.83764 14.5152V8.81965C3.83764 5.87696 5.78523 3.2855 8.63467 2.45016V2.17487C8.63467 1.67136 8.83683 1.18847 9.19667 0.832427C9.55652 0.476388 10.0446 0.276367 10.5535 0.276367C11.0624 0.276367 11.5504 0.476388 11.9103 0.832427C12.2701 1.18847 12.4723 1.67136 12.4723 2.17487V2.45016C15.3217 3.2855 17.2693 5.87696 17.2693 8.81965V14.5152L19.1881 16.4137ZM12.4723 18.3122C12.4723 18.8157 12.2701 19.2986 11.9103 19.6546C11.5504 20.0107 11.0624 20.2107 10.5535 20.2107C10.0446 20.2107 9.55652 20.0107 9.19667 19.6546C8.83683 19.2986 8.63467 18.8157 8.63467 18.3122M17.9889 1.40598L16.6265 2.75392C17.438 3.54694 18.0822 4.49146 18.5219 5.53269C18.9617 6.57393 19.1881 7.69117 19.1881 8.81965H21.1069C21.1069 6.03834 19.994 3.36144 17.9889 1.40598ZM1.8225e-05 8.81965H1.91883C1.91883 6.54144 2.83986 4.35816 4.48044 2.75392L3.11809 1.40598C2.12787 2.37571 1.34229 3.53047 0.807015 4.80318C0.271736 6.07588 -0.00257565 7.44115 1.8225e-05 8.81965Z"
-      fill={color ?? "var(--error-warning-100)"}
-    />
-  </svg>
-);
