--- app/assets/GreenTick.tsx
@@ -1,14 +1,0 @@
-export const GreenTick = () => (
-  <svg
-    width="100%"
-    height="100%"
-    viewBox="0 0 22 18"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M7.12572 18L0 10.8069L3.24731 7.52896L7.12572 11.4556L18.4626 0L21.7099 3.27799L7.12572 18Z"
-      fill="var(--error-success-100)"
-    />
-  </svg>
-);
