--- app/assets/HatIcon.tsx
@@ -1,22 +1,0 @@
-export const HatIcon = ({ color }: { color?: string }) => (
-  <svg
-    width="100%"
-    height="100%"
-    viewBox="0 0 25 18"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M11.9926 0.0755887C12.235 -0.0251962 12.5054 -0.0251962 12.7479 0.0755887L23.4447 4.52265C23.8312 4.68335 24.0847 5.07279 24.0847 5.50588C24.0847 5.93898 23.8312 6.32841 23.4447 6.48912L12.7479 10.9362C12.5054 11.037 12.235 11.037 11.9926 10.9362L1.29582 6.48912C0.909274 6.32841 0.655762 5.93898 0.655762 5.50588C0.655762 5.07279 0.909274 4.68335 1.29582 4.52265L11.9926 0.0755887ZM4.41609 5.50588L12.3702 8.81272L20.3244 5.50588L12.3702 2.19905L4.41609 5.50588Z"
-      fill={color ?? "var(--error-success-100)"}
-    />
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M23.0671 4.50537C23.6292 4.50537 24.0848 4.97942 24.0848 5.56419V10.447C24.0848 11.0317 23.6292 11.5058 23.0671 11.5058C22.5051 11.5058 22.0494 11.0317 22.0494 10.447V5.56419C22.0494 4.97942 22.5051 4.50537 23.0671 4.50537ZM6.53599 6.78978C7.09803 6.78978 7.55366 7.26383 7.55366 7.84861V13.9554C7.57073 13.9699 7.58846 13.9848 7.60687 14.0001C7.85783 14.2089 8.22825 14.4908 8.69473 14.7731C9.63558 15.3424 10.9161 15.8824 12.3703 15.8824C13.8249 15.8824 15.1055 15.3424 16.0464 14.7731C16.5129 14.4908 16.8833 14.2089 17.1343 14.0001C17.1527 13.9848 17.1704 13.9699 17.1875 13.9554V7.84861C17.1875 7.26383 17.6431 6.78978 18.2052 6.78978C18.7672 6.78978 19.2229 7.26383 19.2229 7.84861V14.4355C19.2229 14.7193 19.1134 14.9912 18.919 15.1903L18.2052 14.4355C18.919 15.1903 18.9187 15.1905 18.9184 15.1908L18.9165 15.1927L18.9132 15.1961L18.9038 15.2056C18.8963 15.2131 18.8864 15.2229 18.8741 15.2349C18.8496 15.259 18.8155 15.2917 18.7723 15.332C18.6859 15.4125 18.5626 15.5232 18.4054 15.654C18.0916 15.9151 17.6384 16.2595 17.0696 16.6037C15.9399 17.2872 14.3031 18.0001 12.3703 18.0001C10.4378 18.0001 8.80108 17.2872 7.67146 16.6037C7.10271 16.2595 6.64958 15.9151 6.33579 15.654C6.17854 15.5232 6.05525 15.4125 5.96887 15.332C5.92566 15.2917 5.89159 15.259 5.86705 15.2349C5.85477 15.2229 5.84487 15.2131 5.83738 15.2056L5.82797 15.1961L5.82463 15.1927L5.82273 15.1908C5.82247 15.1905 5.82221 15.1903 6.53599 14.4355L5.82221 15.1903C5.6278 14.9912 5.51831 14.7193 5.51831 14.4355V7.84861C5.51831 7.26383 5.97394 6.78978 6.53599 6.78978Z"
-      fill={color ?? "var(--error-success-100)"}
-    />
-  </svg>
-);
