--- app/assets/Hourglass.tsx
@@ -1,14 +1,0 @@
-export const Hourglass = () => (
-  <svg
-    width="17"
-    height="24"
-    viewBox="0 0 17 24"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M16.8522 20.5898C16.412 16.7303 14.6444 15.1306 13.3534 13.9641C12.4608 13.1544 12.0317 12.7333 12.0317 12C12.0317 11.2767 12.4593 10.8636 13.3494 10.0705C14.6555 8.90748 16.4442 7.3143 16.8532 3.40262C16.8859 3.07399 16.849 2.74217 16.7449 2.42865C16.6407 2.11513 16.4717 1.82688 16.2487 1.58256C16.0106 1.32161 15.7203 1.11326 15.3964 0.97096C15.0725 0.828657 14.7223 0.755553 14.3684 0.75636L3.24746 0.75636C2.89302 0.755257 2.54229 0.828213 2.21789 0.970521C1.89349 1.11283 1.60263 1.32133 1.36409 1.58256C1.14178 1.82722 0.973398 2.11559 0.869807 2.42908C0.766216 2.74257 0.729715 3.07421 0.762657 3.40262C1.17016 7.30175 2.95229 8.88338 4.25337 10.0379C5.15199 10.8355 5.58418 11.2516 5.58418 12C5.58418 12.7579 5.15099 13.1826 4.24934 13.9952C2.96488 15.1547 1.20189 16.7424 0.763665 20.5898C0.728082 20.9169 0.762182 21.2479 0.863737 21.561C0.965292 21.8741 1.13201 22.1623 1.35301 22.4068C1.59205 22.6711 1.88432 22.8822 2.21074 23.0264C2.53716 23.1706 2.89041 23.2446 3.24746 23.2436H14.3684C14.7254 23.2446 15.0787 23.1706 15.4051 23.0264C15.7315 22.8822 16.0238 22.6711 16.2629 22.4068C16.4838 22.1623 16.6506 21.8741 16.7521 21.561C16.8537 21.2479 16.8878 20.9169 16.8522 20.5898ZM13.2053 20.8342H4.43219C3.6464 20.8342 3.42477 19.9307 3.97583 19.3706C5.30966 18.0233 8.00199 17.0586 8.00199 15.5136V10.3937C8.00199 9.39738 6.08789 8.63693 4.90367 7.02066C4.70823 6.75412 4.72787 6.37816 5.22453 6.37816H12.414C12.8376 6.37816 12.9288 6.75111 12.7364 7.01815C11.5693 8.63693 9.61387 9.39236 9.61387 10.3937V15.5136C9.61387 17.046 12.42 17.8727 13.6637 19.3721C14.1649 19.9764 13.9896 20.8342 13.2053 20.8342Z"
-      fill="var(--error-success-100)"
-    />
-  </svg>
-);
