--- app/assets/NotificationIcon.tsx
@@ -1,15 +1,0 @@
-export const NotificationIcon = ({ color }: { color?: string }) => (
-  <svg
-    width="100%"
-    height="100%"
-    viewBox="0 0 19 18"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <rect x="0.5" width="18" height="18" rx="9" fill={color ?? "#D82823"} />
-    <path
-      d="M9.44404 2.44287C9.94558 2.44287 10.349 2.85531 10.338 3.35673L10.1685 11.0615C10.1598 11.4554 9.83804 11.7702 9.44404 11.7702C9.05003 11.7702 8.72823 11.4554 8.71956 11.0615L8.55006 3.35673C8.53903 2.85531 8.94249 2.44287 9.44404 2.44287ZM9.51865 15.5572C9.19531 15.5572 8.9217 15.4452 8.69785 15.2214C8.47399 14.9975 8.36206 14.7239 8.36206 14.4006C8.36206 14.0772 8.47399 13.8036 8.69785 13.5798C8.9217 13.3559 9.19531 13.244 9.51865 13.244C9.82957 13.244 10.0907 13.3559 10.3022 13.5798C10.526 13.8036 10.6379 14.0772 10.6379 14.4006C10.6379 14.7239 10.526 14.9975 10.3022 15.2214C10.0907 15.4452 9.82957 15.5572 9.51865 15.5572Z"
-      fill="white"
-    />
-  </svg>
-);
