--- app/assets/PurpleStar.tsx
@@ -1,14 +1,0 @@
-export const PurpleStar = () => (
-  <svg
-    width="100%"
-    height="100%"
-    viewBox="0 0 19 18"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M4.00096 17.7714L5.48829 11.2852L0.5 6.92253L7.09004 6.34546L9.65283 0.228516L12.2156 6.34546L18.8057 6.92253L13.8174 11.2852L15.3047 17.7714L9.65283 14.3321L4.00096 17.7714Z"
-      fill="var(--theme-module3-100)"
-    />
-  </svg>
-);
