--- app/assets/RedCross.tsx
@@ -1,14 +1,0 @@
-export const RedCross = () => (
-  <svg
-    width="100%"
-    height="100%"
-    viewBox="0 0 18 18"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M15.6959 0.405762L9.24924 6.85144L2.80413 0.405762L0.655762 2.55432L7.10087 9L0.655762 15.4457L2.80413 17.5942L9.24924 11.1486L15.6959 17.5942L17.8442 15.4457L11.3991 9L17.8442 2.55432L15.6959 0.405762Z"
-      fill="var(--error-failure-100)"
-    />
-  </svg>
-);
