+++ app/assets/favicon.ico
