+++ app/components/statuses/statuses.tsx
@@ -5,11 +5,13 @@ import {
   ReturnStatus,
 } from "../../guides/types";
 import { Grade, IconContainer, Status, StatusesWrapper } from "./style";
-import { GreenTick } from "../../assets/greenTick";
-import { PurpleStar } from "../../assets/PurpleStar";
-import { RedCross } from "../../assets/RedCross";
-import { Bell } from "../../assets/Bell";
-import { Hourglass } from "../../assets/Hourglass";
+import {
+  Bell,
+  GreenTick,
+  PurpleStar,
+  RedCross,
+  Hourglass,
+} from "../../assets/Icons";
 
 export const Statuses = ({
   returnStatus,
