+++ app/globalStyles/buttons/exit/index.tsx
@@ -1,17 +1,16 @@
-import { Button, ExitIcon } from './style';
-import exiticon from '../../../assets/exit.svg';
+import { Button } from "./style";
+import { ExitIcon } from "../../../assets/Icons";
 
 type Props = {
-    onClick: () => void;
+  onClick: () => void;
+};
 
-}
+const ExitButton = ({ onClick }: Props) => {
+  return (
+    <Button onClick={onClick}>
+      <ExitIcon />
+    </Button>
+  );
+};
 
-const ExitButton = ({onClick}: Props) => {
-    return ( 
-        <Button onClick={onClick}>
-            <ExitIcon src={exiticon} alt="exit" />
-        </Button>
-     );
-}
- 
 export default ExitButton;
